+++ LibraryManagementSystem/Controllers/FeedbacksController.cs
@@ -15,10 +15,25 @@ namespace LibraryManagementSystem.Controllers
     {
         private LibraryManagementSystemContext db = new LibraryManagementSystemContext();
 
+        // SECURITY CHECK: Login required
+        protected override void OnActionExecuting(ActionExecutingContext filterContext)
+        {
+            if (Session["UserName"] == null)
+            {
+                filterContext.Result = RedirectToAction("Login", "Home");
+                return;
+            }
+
+            base.OnActionExecuting(filterContext);
+        }
+
         // GET: Feedbacks
         public ActionResult Index()
         {
-            var feedbacks = db.Feedbacks.Include(f => f.Book).Include(f => f.Member);
+            var feedbacks = db.Feedbacks
+                .Include(f => f.Book)
+                .Include(f => f.Member);
+
             return View(feedbacks.ToList());
         }
 
@@ -29,11 +44,14 @@ namespace LibraryManagementSystem.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             Feedback feedback = db.Feedbacks.Find(id);
+
             if (feedback == null)
             {
                 return HttpNotFound();
             }
+
             return View(feedback);
         }
 
@@ -42,12 +60,11 @@ namespace LibraryManagementSystem.Controllers
         {
             ViewBag.BookId = new SelectList(db.Books, "Id", "Title");
             ViewBag.MemberId = new SelectList(db.Members, "Id", "Name");
+
             return View();
         }
 
         // POST: Feedbacks/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,MemberId,BookId,Rating,Comment")] Feedback feedback)
@@ -56,11 +73,13 @@ namespace LibraryManagementSystem.Controllers
             {
                 db.Feedbacks.Add(feedback);
                 db.SaveChanges();
+
                 return RedirectToAction("Index");
             }
 
             ViewBag.BookId = new SelectList(db.Books, "Id", "Title", feedback.BookId);
             ViewBag.MemberId = new SelectList(db.Members, "Id", "Name", feedback.MemberId);
+
             return View(feedback);
         }
 
@@ -71,19 +90,21 @@ namespace LibraryManagementSystem.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             Feedback feedback = db.Feedbacks.Find(id);
+
             if (feedback == null)
             {
                 return HttpNotFound();
             }
+
             ViewBag.BookId = new SelectList(db.Books, "Id", "Title", feedback.BookId);
             ViewBag.MemberId = new SelectList(db.Members, "Id", "Name", feedback.MemberId);
+
             return View(feedback);
         }
 
         // POST: Feedbacks/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,MemberId,BookId,Rating,Comment")] Feedback feedback)
@@ -92,10 +113,13 @@ namespace LibraryManagementSystem.Controllers
             {
                 db.Entry(feedback).State = EntityState.Modified;
                 db.SaveChanges();
+
                 return RedirectToAction("Index");
             }
+
             ViewBag.BookId = new SelectList(db.Books, "Id", "Title", feedback.BookId);
             ViewBag.MemberId = new SelectList(db.Members, "Id", "Name", feedback.MemberId);
+
             return View(feedback);
         }
 
@@ -106,11 +130,14 @@ namespace LibraryManagementSystem.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             Feedback feedback = db.Feedbacks.Find(id);
+
             if (feedback == null)
             {
                 return HttpNotFound();
             }
+
             return View(feedback);
         }
 
@@ -120,8 +147,13 @@ namespace LibraryManagementSystem.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Feedback feedback = db.Feedbacks.Find(id);
-            db.Feedbacks.Remove(feedback);
-            db.SaveChanges();
+
+            if (feedback != null)
+            {
+                db.Feedbacks.Remove(feedback);
+                db.SaveChanges();
+            }
+
             return RedirectToAction("Index");
         }
 
@@ -131,6 +163,7 @@ namespace LibraryManagementSystem.Controllers
             {
                 db.Dispose();
             }
+
             base.Dispose(disposing);
         }
     }
